services/api/pyproject.toml
[project]
name = "co-op-api"
version = "0.2.0"
description = "FastAPI Main Control Plane for Co-Op AI OS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "sqlalchemy>=2.0.30",
    "asyncpg>=0.29.0",
    "pydantic-settings>=2.2.1",
    "redis>=5.0.4",
    "qdrant-client>=1.9.1",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "structlog>=24.2.0",
    "asgi-correlation-id>=4.3.1",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "slowapi>=0.1.9",
    "minio>=7.2.5",
    "python-multipart>=0.0.9",
    "aiofiles>=23.2.1",
    "pymupdf>=1.24.1",
    "torch>=2.0.0",
    "sentence-transformers>=2.6.1",
    "langchain-core>=0.1.40",
    "langgraph>=0.0.34",
    "sse-starlette>=2.1.0",
    "python-docx>=1.1.0",
    # --- Stage 2 additions ---
    "arq>=0.26.1",
    "python-telegram-bot>=21.0",
    "psutil>=5.9.8",
    "py-cpuinfo>=9.0.0",
    # SECURITY (CVE-2026-33634): LiteLLM pinned to ==1.82.6.
    # Versions 1.82.7 and 1.82.8 were compromised with credential-stealing
    # malware on March 24, 2026. Do NOT change this without verifying the
    # new version does not contain litellm_init.pth.
    "litellm==1.82.6",
    "alembic>=1.13.1",
    "sentry-sdk[fastapi]>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "ruff>=0.4.8",
    "mypy>=1.10.0",
    "pip-audit>=2.7.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
